next-backup-to-remove.awk: Fix bi-weekly year change
Using integer division by 2 to determine if two weeks are adjacent does
not work around a year change. Bi-weeks 26 and 0 can represent adjacent
weeks, but can be more than a week apart as well.

This commit adds a mechanism to detect if week 0 (the first week of the
year) is next to the week of the last backup from the previous year. If
this is the case, the week 0 backups are daily backups.