Added detection of mysqldump --events to global conf.
MySQL added so called events, which are not automatically backed up
by mysqldump. If mysqldump supports dumping of events, it provides
the command line switch --events to do so. If --events is supported
but not specified, a warning is displayed.

This commit adds code to automatically add --events if supported.
A complete backup should include the events.

Adding the --events option can be overridden in /etc/fgn-backup.conf
by specifiying the wanted options in the MYSQLDUMPFLAGS variable.